enhancement: Add `cerbos.epdp.v2` messages (#348)
I realised https://github.com/cerbos/cerbos/pull/2770 was probably a
mistake and the `Config` message more logically belongs in the
`cerbos.epdp` namespace. This PR moves it from `cerbos.engine.v1` to
`cerbos.epdp.v2` and adds `Metadata` to be included in the build
artifacts and an `Error` message to replace the current approach of
returning a custom JSON structure on errors.

Signed-off-by: Andrew Haines <haines@cerbos.dev>